+++ types/index.d.ts
@@ -22,6 +22,7 @@ export {
 } from './insert';
 export {default as intersection} from './intersection';
 export {default as isSubsetOf} from './isSubsetOf';
+export {default as isSupersetOf} from './isSupersetOf';
 export {default as iterate} from './iterate';
 export {default as join} from './join';
 export {default as join2} from './join2';