+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
   * Added `slice`, which works like `Array.prototype.slice`.
   * Added `isSubsetOf`, which checks if `tree` is a subset of `other`.
+  * Added `isSupersetOf`, which checks if `tree` is a superset of `other`.
   * Fixed the return type of `findNode` to use `NonEmptyImmutableTree`.
 
 ## v0.12.0 (2025-07-10)